feat: move react wrappers to own package
