+++ react/HotButton.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { createComponent } from '@lit/react';
-import Button from '../../../components/button/Button';
+import Button from '../components/button/Button';
 
 
 const reactWrapper = createComponent({
